# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão de Tarefas",
        "description": "Documentação da API de Gestão de Tarefas com Laravel Sanctum",
        "contact": {
            "email": "seu_email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Servidor da API"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/user/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registra um novo usuário",
                "operationId": "1f06d3130b898773356c61c7e17c857a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Daniel"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "daniel@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso"
                    },
                    "500": {
                        "description": "Erro ao registrar usuário"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autentica um usuário e retorna o token JWT",
                "operationId": "72fe6885a7a3aa143f006083634ba4cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "daniel@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Termina a sessão do usuário autenticado",
                "operationId": "0220c7a237541f2dfc8ac465999dbfe6",
                "responses": {
                    "200": {
                        "description": "Logout bem-sucedido"
                    },
                    "401": {
                        "description": "Falha no logout"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Listar todas as tarefas do usuário autenticado",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "responses": {
                    "200": {
                        "description": "Lista de tarefas retornada com sucesso"
                    },
                    "404": {
                        "description": "Nenhuma tarefa encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Criar nova tarefa",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Comprar pão"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Ir à padaria comprar pão às 7h"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarefa criada com sucesso"
                    },
                    "500": {
                        "description": "Falha ao criar tarefa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Buscar tarefa pelo ID",
                "operationId": "fd3677b5b24695d7b801715869b1357c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa encontrada"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Atualizar uma tarefa existente",
                "operationId": "85377ff0e935f093dd75ea6886923dbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Atualizar tarefa"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nova descrição da tarefa"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed"
                                        ],
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada com sucesso"
                    },
                    "500": {
                        "description": "Erro ao atualizar tarefa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Excluir uma tarefa pelo ID",
                "operationId": "5ed5dfd4c645e9398f4282969bfe6478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa excluída com sucesso"
                    },
                    "500": {
                        "description": "Erro ao excluir tarefa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/status/{status}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Filtrar tarefas por status",
                "operationId": "84f21b4f75d2afd448278a6fd92a2046",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Status da tarefa (pending ou completed)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "completed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefas com o status especificado"
                    },
                    "404": {
                        "description": "Nenhuma tarefa com esse status"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Insere o token JWT no formato: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Endpoints para autenticação de usuários"
        },
        {
            "name": "Tasks",
            "description": "Tasks management endpoints"
        }
    ]
}